finish first draft of calculator with functional operator buttons

+++ calculatormainwindow.h
@@ -31,7 +31,7 @@ class CalculatorMainWindow : public QMainWindow {
 
    private:
    Ui::CalculatorMainWindow *ui;
-   int currentValue;
+   double currentValue;
    char currentOperator;
 };
 #endif // CALCULATORMAINWINDOW_H
